Se hizo una validacion en las alarmas de Binance para filtrar que sea numerico.

+++ alarm.php
@@ -17,15 +17,18 @@
 		$type = $row['type'];
 		$row_num = $row['row_num'];
 		$price = json_decode(file_get_contents("https://api.binance.com/api/v1/ticker/price?symbol=$coin"), true)['price'];
-		$seted_price = floatval($seted_price);
-		$price = floatval($price);
-		if ($price >= $seted_price and $type == "high") {
-			sendMessage($chat_id, $coin." just reached the price of ".$seted_price);
-			mysqli_query($link, "DELETE FROM alarms_binance WHERE row_num ='$row_num'");
-		}
-		elseif ($price <= $seted_price and $type == "low") {
-			sendMessage($chat_id, "/".$coin." just reached the price of ".$seted_price);
-			mysqli_query($link, "DELETE FROM alarms_binance WHERE row_num ='$row_num'");
+		if (!is_numeric($price)) {
+			# code...
+			$seted_price = floatval($seted_price);
+			$price = floatval($price);
+			if ($price >= $seted_price and $type == "high") {
+				sendMessage($chat_id, $coin." just reached the price of ".$seted_price);
+				mysqli_query($link, "DELETE FROM alarms_binance WHERE row_num ='$row_num'");
+			}
+			elseif ($price <= $seted_price and $type == "low") {
+				sendMessage($chat_id, "/".$coin." just reached the price of ".$seted_price);
+				mysqli_query($link, "DELETE FROM alarms_binance WHERE row_num ='$row_num'");
+			}
 		}
 
  	}
